s01e05/requirements.txt
python-dotenv
requests
langchain
langchain-openai
langgraph
